Add Support for Java with Jbang
Jbang was needed because the ObjectMapper jar is needed
Currently it works, but I think there could be another way
to do this without the use of JBang, using only Java.
And Placing the JAR on the classpath.
Since JBang builds the jar everytime before execution
that also causes delay and can result in test failures dut to timeout.
But at the same time JBang opens up for easier customization,
and support for Kotlin/Groovy yoo, so that is a tradeoff.

JBang catalog/aliases can help with the rebuild issue
but on first run/test timeout might be an issue.
Timeout might also depends on the machine specs.